eliminate validations that were done twice
