+++ src/main/scala/nl.knaw.dans.easy.multideposit/parser/FileMetadataParser.scala
@@ -27,14 +27,10 @@ trait FileMetadataParser extends DebugEnhancedLogging {
   def extractFileMetadata(depositDir: File, instructions: Instructions): Try[Seq[FileMetadata]] = {
     logger.info(s"extract metadata from files in $depositDir")
     if (depositDir.exists)
-      for {
-        fms <- depositDir.listRecursively.filterNot(_.isDirectory)
-          .map(getFileMetadata(instructions))
-          .toList
-          .collectResults
-        _ <- checkSFColumnsIfDepositContainsAVFiles(instructions, fms)
-        _ <- checkEitherVideoOrAudio(fms, instructions.row)
-      } yield fms
+      depositDir.listRecursively
+        .collect { case file if !file.isDirectory => getFileMetadata(instructions)(file) }
+        .toList
+        .collectResults
     else // if the deposit does not contain any data
       Success { List.empty }
   }
@@ -120,31 +116,5 @@ trait FileMetadataParser extends DebugEnhancedLogging {
 
   private def defaultAccessibility(instructions: Instructions): FileAccessRights.Value = {
     FileAccessRights.accessibleTo(instructions.profile.accessright)
-  }
-
-  private def checkSFColumnsIfDepositContainsAVFiles(instructions: Instructions,
-                                                     fileMetadata: Seq[FileMetadata]): Try[Unit] = {
-    val avFiles = fileMetadata.collect { case fmd: AVFileMetadata => fmd.filepath }
-
-    (instructions.audioVideo.springfield.isDefined, avFiles.isEmpty) match {
-      case (true, false) | (false, true) => Success(())
-      case (true, true) =>
-        Failure(ParseException(instructions.row,
-          "Values found for these columns: [SF_DOMAIN, SF_USER, SF_COLLECTION, SF_PLAY_MODE]; " +
-            "these columns should be empty because there are no audio/video files " +
-            "found in this deposit"))
-      case (false, false) =>
-        Failure(ParseException(instructions.row,
-          "No values found for these columns: [SF_USER, SF_COLLECTION, SF_PLAY_MODE]; " +
-            "these columns should contain values because audio/video files are " +
-            s"found:\n${ avFiles.map(filepath => s" - $filepath").mkString("\n") }"))
-    }
-  }
-
-  private def checkEitherVideoOrAudio(fileMetadata: Seq[FileMetadata], row: Int): Try[Unit] = {
-    fileMetadata.collect { case fmd: AVFileMetadata => fmd.vocabulary }.distinct match {
-      case Nil | Seq(_) => Success(())
-      case _ => Failure(ParseException(row, "Found both audio and video in this dataset. Only one of them is allowed."))
-    }
   }
 }
