+++ src/main/scala/nl.knaw.dans.easy.multideposit/parser/Validation.scala
@@ -1,0 +1,60 @@
+package nl.knaw.dans.easy.multideposit.parser
+
+import nl.knaw.dans.easy.multideposit.model.{ AVFileMetadata, Deposit }
+import nl.knaw.dans.lib.logging.DebugEnhancedLogging
+
+import scala.util.{ Failure, Success, Try }
+
+trait Validation extends DebugEnhancedLogging {
+
+  def validateDeposit(deposit: Deposit): Try[Unit] = {
+    logger.debug(s"validating deposit ${ deposit.depositId }")
+
+    for {
+      _ <- checkSpringFieldDepositHasAVformat(deposit)
+      _ <- checkSFColumnsIfDepositContainsAVFiles(deposit)
+      _ <- checkEitherVideoOrAudio(deposit)
+    } yield ()
+  }
+
+  def checkSpringFieldDepositHasAVformat(deposit: Deposit): Try[Unit] = {
+    logger.debug("check that a Springfield deposit has an A/V format")
+
+    deposit.springfield match {
+      case None => Success(())
+      case Some(_) => deposit.metadata.formats
+        .find(s => s.startsWith("audio/") || s.startsWith("video/"))
+        .map(_ => Success(()))
+        .getOrElse(Failure(ParseException(deposit.row,
+          "No audio/video format found for this column: [DC_FORMAT]\n" +
+            "cause: this column should contain at least one " +
+            "audio/ or video/ value because SF columns are present")))
+    }
+  }
+
+  def checkSFColumnsIfDepositContainsAVFiles(deposit: Deposit): Try[Unit] = {
+    val avFiles = deposit.files.collect { case fmd: AVFileMetadata => fmd.filepath }
+
+    (deposit.springfield.isDefined, avFiles.isEmpty) match {
+      case (true, false) | (false, true) => Success(())
+      case (true, true) =>
+        Failure(ParseException(deposit.row,
+          "Values found for these columns: [SF_DOMAIN, SF_USER, SF_COLLECTION, SF_PLAY_MODE]; " +
+            "these columns should be empty because there are no audio/video files " +
+            "found in this deposit"))
+      case (false, false) =>
+        Failure(ParseException(deposit.row,
+          "No values found for these columns: [SF_USER, SF_COLLECTION, SF_PLAY_MODE]; " +
+            "these columns should contain values because audio/video files are " +
+            s"found:\n${ avFiles.map(filepath => s" - $filepath").mkString("\n") }"))
+    }
+  }
+
+  def checkEitherVideoOrAudio(deposit: Deposit): Try[Unit] = {
+    deposit.files.collect { case fmd: AVFileMetadata => fmd.vocabulary }.distinct match {
+      case Nil | Seq(_) => Success(())
+      case _ => Failure(ParseException(deposit.row,
+        "Found both audio and video in this dataset. Only one of them is allowed."))
+    }
+  }
+}
