+++ src/test/scala/nl.knaw.dans.easy.multideposit/actions/ValidatePreconditionsSpec.scala
@@ -16,9 +16,8 @@
 package nl.knaw.dans.easy.multideposit.actions
 
 import better.files.File
-import better.files.File.currentWorkingDirectory
 import javax.naming.directory.Attributes
-import nl.knaw.dans.easy.multideposit.model.{ AVFileMetadata, Audio, FileAccessRights, Metadata, PlayMode, Springfield, Video }
+import nl.knaw.dans.easy.multideposit.model.{ AVFileMetadata, FileAccessRights, Video }
 import nl.knaw.dans.easy.multideposit.{ FfprobeRunner, Ldap, TestSupportFixture }
 import nl.knaw.dans.lib.error.CompositeException
 import org.scalamock.scalatest.MockFactory
@@ -67,29 +66,6 @@ class ValidatePreconditionsSpec extends TestSupportFixture with BeforeAndAfterEa
     }
   }
 
-  "checkSpringFieldDepositHasAVformat" should "fail if the deposit contains SF_* fields, but no AV DC_FORMAT is given" in {
-    val deposit = testInstructions1.toDeposit().copy(
-      depositId = depositId,
-      metadata = Metadata(
-        formats = List("text/plain")
-      )
-    )
-    inside(action.checkSpringFieldDepositHasAVformat(deposit)) {
-      case Failure(InvalidInputException(_, message)) =>
-        message should include("No audio/video format found for this column: [DC_FORMAT]")
-    }
-  }
-
-  it should "succeed if the deposit contains SF_* fields, and the DC_FORMAT contains audio/" in {
-    val deposit = testInstructions1.toDeposit().copy(
-      depositId = depositId,
-      metadata = Metadata(
-        formats = List("audio/mpeg3")
-      )
-    )
-    action.checkSpringFieldDepositHasAVformat(deposit) shouldBe a[Success[_]]
-  }
-
   val avFileReferences = Seq(
     AVFileMetadata(
       filepath = testDir / "md" / "ruimtereis01" / "reisverslag" / "centaur.mpg",
@@ -99,75 +75,6 @@ class ValidatePreconditionsSpec extends TestSupportFixture with BeforeAndAfterEa
       accessibleTo = FileAccessRights.ANONYMOUS,
       visibleTo = FileAccessRights.ANONYMOUS
     ))
-
-  "checkSFColumnsIfDepositContainsAVFiles" should "succeed if the deposit contains the SF_* fields in case an A/V file is found" in {
-    val deposit = testInstructions1.toDeposit(avFileReferences).copy(
-      depositId = depositId,
-      springfield = Option(Springfield("domain", "user", "collection", PlayMode.Continuous))
-    )
-    action.checkSFColumnsIfDepositContainsAVFiles(deposit) shouldBe a[Success[_]]
-  }
-
-  it should "fail if the deposit contains A/V files but the SF_* fields are not present" in {
-    val deposit = testInstructions1.toDeposit(avFileReferences).copy(
-      depositId = depositId,
-      springfield = Option.empty
-    )
-    inside(action.checkSFColumnsIfDepositContainsAVFiles(deposit)) {
-      case Failure(InvalidInputException(_, message)) =>
-        message should {
-          include("No values found for these columns: [SF_USER, SF_COLLECTION, SF_PLAY_MODE]") and
-            include("reisverslag/centaur.mpg")
-        }
-    }
-  }
-
-  it should "succeed if the deposit contains no A/V files and the SF_* fields are not present" in {
-    val depositId = "ruimtereis02"
-    val deposit = testInstructions2.toDeposit().copy(
-      depositId = depositId,
-      springfield = Option.empty
-    )
-    action.checkSFColumnsIfDepositContainsAVFiles(deposit) shouldBe a[Success[_]]
-  }
-
-  it should "fail if the deposit contains no A/V files and any of the SF_* fields are present" in {
-    val depositId = "ruimtereis02"
-    val deposit = testInstructions2.toDeposit().copy(
-      row = 1,
-      depositId = depositId,
-      springfield = Option(Springfield(user = "user", collection = "collection", playMode = PlayMode.Continuous))
-    )
-    inside(action.checkSFColumnsIfDepositContainsAVFiles(deposit)) {
-      case Failure(InvalidInputException(_, message)) =>
-        message should include("Values found for these columns: [SF_DOMAIN, SF_USER, SF_COLLECTION, SF_PLAY_MODE]; these columns should be empty because there are no audio/video files found in this deposit")
-    }
-  }
-
-  it should "create an empty list of file metadata if the deposit directory corresponding with the depositId does not exist and therefore succeed" in {
-    val depositId = "ruimtereis03"
-    val deposit = testInstructions2.copy(depositId = depositId).toDeposit()
-    depositDir(depositId).toJava should not(exist)
-    action.checkSFColumnsIfDepositContainsAVFiles(deposit) shouldBe a[Success[_]]
-  }
-
-  "checkEitherVideoOrAudio" should "fail if a dataset has both audio and video material in it" in {
-    val depositId = "ruimtereis01"
-    val deposit = testInstructions1.copy(depositId = depositId)
-      .toDeposit(avFileReferences :+ AVFileMetadata(
-        filepath = currentWorkingDirectory,
-        mimeType = "audio/mpeg",
-        vocabulary = Audio,
-        title = "mytitle",
-        accessibleTo = FileAccessRights.ANONYMOUS,
-        visibleTo = FileAccessRights.ANONYMOUS
-      ))
-
-    inside(action.checkEitherVideoOrAudio(deposit)) {
-      case Failure(InvalidInputException(_, message)) =>
-        message shouldBe "Found both audio and video in this dataset. Only one of them is allowed."
-    }
-  }
 
   def mockLdapForDepositor(expectedResult: Try[Seq[Boolean]]): Unit = {
     (ldapMock.query(_: String)(_: Attributes => Boolean)) expects("dp1", *) returning expectedResult
