+++ src/test/scala/nl.knaw.dans.easy.multideposit/parser/FileMetadataParserSpec.scala
@@ -184,55 +184,4 @@ class FileMetadataParserSpec extends TestSupportFixture with FileMetadataTestObj
         e1 should have message s"Not listed A/V file detected: $fileWithNoDescription. Because Springfield PlayMode 'MENU' was choosen, all A/V files must be listed with a human readable title in the FILE_TITLE field."
     }
   }
-
-  it should "fail if the deposit contains A/V files but the SF_* fields are not present" in {
-    val instructions = testInstructions1.copy(
-      depositId = "ruimtereis01",
-      audioVideo = AudioVideo(springfield = Option.empty)
-    )
-    inside(extractFileMetadata(multiDepositDir / "ruimtereis01", instructions)) {
-      case Failure(ParseException(_, message, _)) =>
-        message should {
-          include("No values found for these columns: [SF_USER, SF_COLLECTION, SF_PLAY_MODE]") and
-            include("reisverslag/centaur.mpg") and
-            include("path/to/a/random/video/hubble.mpg")
-        }
-    }
-  }
-
-  it should "fail if the deposit contains no A/V files but any of the SF_* fields are not present" in {
-    val instructions = testInstructions2.copy(
-      depositId = "ruimtereis02",
-      audioVideo = testInstructions2.audioVideo.copy(
-        springfield = Option(Springfield(user = "user", collection = "collection", playMode = PlayMode.Continuous))
-      )
-    )
-    inside(extractFileMetadata(multiDepositDir / "ruimtereis02", instructions)) {
-      case Failure(ParseException(_, message, _)) =>
-        message should {
-          include("Values found for these columns: [SF_DOMAIN, SF_USER, SF_COLLECTION, SF_PLAY_MODE]") and
-            include("these columns should be empty because there are no audio/video files found in this deposit")
-        }
-    }
-  }
-
-  it should "fail if a dataset has both audio and video material in it" in {
-    val instructions = testInstructions1.copy(depositId = "ruimtereis01")
-
-    val audioFile = multiDepositDir / "ruimtereis01/path/to/a/random/audio/chicken.mp3"
-    audioFile.parent.createDirectories()
-    (multiDepositDir / "ruimtereis04/path/to/a/random/sound/chicken.mp3").copyTo(audioFile)
-
-    val currentAV = instructions.audioVideo.avFiles
-    val newAV = currentAV + (audioFile -> Set.empty[Subtitles])
-    val failingInstructions = instructions.copy(
-      audioVideo = instructions.audioVideo.copy(avFiles = newAV),
-      files = instructions.files + (audioFile -> FileDescriptor(title = Option("crowing rooster")))
-    )
-
-    inside(extractFileMetadata(multiDepositDir / "ruimtereis01", failingInstructions)) {
-      case Failure(ParseException(_, message, _)) =>
-        message shouldBe "Found both audio and video in this dataset. Only one of them is allowed."
-    }
-  }
 }
